fix failing tests

+++ python/demos/demo_christmas_tree.py
@@ -142,7 +142,7 @@ if __name__ == "__main__":
 
     for i in range(split):
         fcts = np.array(find_candidate_surface_segment(
-            mesh, facet_marker.find(marker_offset + split + i), cand_facets_0, 0.8), dtype=np.int32)
+            mesh, facet_marker.find(marker_offset + split + i), cand_facets_0, 0.8), dtype=np.int64)
         vls = np.full(len(fcts), marker_offset + 2 * split + i, dtype=np.int32)
         mts.append(meshtags(mesh, tdim - 1, fcts, vls))
